tests/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pywin32>=305
winocr>=0.0.15
pytest>=7.0.0
pytest-timeout>=2.0.0
